+++ src/main/java/web/dao/UserDAOImpl.java
@@ -27,19 +27,19 @@ public class UserDAOImpl implements UserDAO {
 
     @Override
     @Transactional
-    public User show(int id) {
+    public User show(Long id) {
         return entityManager.find(User.class, id);
     }
 
     @Override
     @Transactional
-    public void update(int id, User updatedUser) {
+    public void update(Long id, User updatedUser) {
         entityManager.merge(updatedUser);
     }
 
     @Override
     @Transactional
-    public void delete(int id) {
+    public void delete(Long id) {
         entityManager.remove(entityManager.find(User.class, id));
     }
 }
